pyproject.toml
[tool.poetry]
name = "template_for_fastapi"
version = "0.1.0"
description = "Template for fastapi, sqlalchemy, pydantic tools"
authors = ["kolya <fr13ndyy@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.83.0"
SQLAlchemy = "^1.4.41"
pydantic = "^1.10.2"
asyncpg = "^0.26.0"
psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"
alembic = "^1.8.1"
pytest = "^7.1.3"
httpx = "^0.23.0"
SQLAlchemy-Utils = "^0.38.3"
pylint = "^2.15.2"
isort = "^5.10.1"
black = "^22.8.0"
ConfigArgParse = "^1.5.3"
pytest-asyncio = "^0.19.0"
uvicorn = "^0.18.3"

[tool.poetry.dev-dependencies]

# Pytest
# ---------------
[tool.pytest]
python_files = "test_*.py"
python_classes = ["*Test", "Test*"]
python_functions = "test_*"
testpaths = "tests/"

[tool.pytest.ini_options]
asyncio_mode = "auto"
# ---------------

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

